Change check_all_sources() from a periodic job to a management command.
The once-a-day job seems to be greatly straining resources at CoralNet's current scale, and the logic to trigger checks outside of this daily job has probably gotten smart enough to stand mostly on its own. For the cases where it misses, we can add a manual button.
