+++ project/vision_backend/tests/tasks/test_misc.py
@@ -7,9 +7,7 @@ from annotations.models import Annotation
 from jobs.models import Job
 from jobs.tasks import run_scheduled_jobs_until_empty
 from jobs.utils import schedule_job
-from lib.tests.utils import ClientTest
 from ...models import Score, Classifier
-from ...tasks import check_all_sources
 from .utils import BaseTaskTest, do_collect_spacer_jobs
 
 
@@ -216,39 +214,3 @@ class CollectSpacerJobsTest(BaseTaskTest):
         self.assertEqual(
             Job.objects.filter(job_name='collect_spacer_jobs').count(), 1,
             "Should not have accepted the second run")
-
-
-class CheckAllSourcesTest(ClientTest):
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-        cls.user = cls.create_user()
-        cls.source = cls.create_source(cls.user)
-        cls.source2 = cls.create_source(cls.user)
-
-    @staticmethod
-    def run_and_get_result():
-        # Note that this may or may not schedule a new job instance; perhaps
-        # the periodic job was already scheduled at the end of the previous
-        # job's run.
-        schedule_job('check_all_sources')
-        check_all_sources()
-        job = Job.objects.filter(
-            job_name='check_all_sources',
-            status=Job.Status.SUCCESS).latest('pk')
-        return job.result_message
-
-    def test(self):
-        self.assertEqual(
-            self.run_and_get_result(),
-            "Scheduled checks for 2 source(s)")
-
-        # If these lines don't get errors, then the expected
-        # scheduled jobs exist
-        Job.objects.get(
-            job_name='check_source', arg_identifier=self.source.pk,
-            status=Job.Status.PENDING)
-        Job.objects.get(
-            job_name='check_source', arg_identifier=self.source2.pk,
-            status=Job.Status.PENDING)
